chore: Update dependencies and add AIChat components for chat functionality
