+++ app/page.tsx
@@ -2,7 +2,7 @@
 import { headers } from 'next/headers'
 import { getDictionary } from './dictionaries'
 import PageContent from '@/components/PageContent'
-import AIChatButton from '@/components/AIChat/AIChatButton'
+import AIChat from '@/components/AIChat'
 
 const locales = ['pt-BR', 'en-US']
 
@@ -13,7 +13,7 @@ const Home = async () => {
   return (
     <main className="relative mx-auto flex flex-col items-center justify-center overflow-hidden bg-black-100 px-5 sm:px-10">
       <PageContent dictionary={dictionary} />
-      <AIChatButton />
+      <AIChat />
     </main>
   )
 }
